backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
supabase
alpaca-py
pydantic>=2
asyncpg
jsonschema
yfinance
pandas
